pyproject.toml
[project]
name = "tc25-ddq-hot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.1",
    "tableauserverclient>=0.36",
]

[dependency-groups]
jwt = [
    "pyjwt>=2.10.1",
]
